[fix] thievescount < 0 occurs, cast for signed comparison

+++ src/task/kaapi_task_finalize.c
@@ -55,7 +55,8 @@ void kaapi_taskfinalize_body( void* taskarg, kaapi_thread_t* thread )
   while (KAAPI_ATOMIC_READ(&ta->sc.is_there_thief) !=0)
     kaapi_slowdown_cpu();
 
-  while (KAAPI_ATOMIC_READ( &ta->thievescount ) >0)
+  /* sometimes drops to -1... */
+  while ((int)KAAPI_ATOMIC_READ( &ta->thievescount ) >0)
     kaapi_slowdown_cpu();
 
   kaapi_readmem_barrier(); /* avoid read reorder before the barrier, for instance reading some data */
